less magical child pinatas
Return null instead of returning a failed pinata, this avoids confusion
when calling ClearError() on a failed child pinata and is also more
consistent with all other methods (returning default value upon
failure). It also avoids an allocation.

The disadvantage of this is that chaining calls for child pinatas is no
longer possible.